Add a mailer report view
This is useful for support staff, administrators, developers, etc. to get a
quick heads-up as to how the mailer is working. Obviously it can't answer
every question, but it should make it obvious if the mailer is failing to
send messages or is falling behind.

To enable in your project, do the normal include('mailer.urls') trick in
your root urls mapping to somewhere sane like r"^mailer/". The view itself
is protected by a permission_required decorator; to view it a user must have
the "mailer.add_message" permission.

Signed-off-by: Dan McGee <dpmcgee@gmail.com>