Add ability to get arbitrary percentiles of delay values
A nice little hack that lets us pass values from the template to get data we
might want there. This will allow anyone to get as much or as few values as
they might need. There is also a coercion of the '_' character into a '.'
so percentiles such as '99.9' can be obtained.

Signed-off-by: Dan McGee <dpmcgee@gmail.com>