Base logging level on verbosity setting
Whenever you call a management command, you can specify a verbosity. Use
this to determine what level to log at so cron jobs can be silent most of
the time, but will still spew output when necessary. This also moves the
default logging level from debug to info, but the old loggers can be seen by
passing '-v 2' to the send_mail and retry_deferred commands.

Signed-off-by: Dan McGee <dpmcgee@gmail.com>